Put text frag logic into text app
