+++ api/text/admin.py
@@ -1,5 +1,7 @@
 from django.contrib import admin
-from text.models import Text
+from text.models.text import Text
+from text.models.fragment import Fragment
 
 # Register your models here.
 admin.site.register(Text)
+admin.site.register(Fragment)
