--- api/text_frag/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from text_frag.models import TextFrag
-
-# Register your models here.
-
-admin.site.register(TextFrag)
