--- api/text_frag/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class TextFragConfig(AppConfig):
-    name = 'text_frag'
