--- api/text_frag/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from text_frag.views import TextFragView
-
-urlpatterns = [
-    path('', TextFragView.as_view()),
-]
